requirements.txt
requests==2.26.0
cryptography==36.0.1
python-dotenv==0.19.2
